UI/Style changes to pages, higher functionality on user details page
These are attempts to create a unified UI for the extant pages in the base folder.

Thymeleaf is not fully integrated. I have attempted to make notes of where further work needs to be done on integration, specifically on userDetails.html. For the time being, most of the functionality is partly there but hidden so as to not obstruct the demo info to show how the pages look.