Additional Pages
These are additional templates for remaining pages for 'Trip Details' (a public facing page where other users can see a trip), and 'Edit Trip Details' which is meant as a private user page, in addition to a very barebones mockup of how the search page should be laid out. As a note, in lieu of adding experiences directly via trip details, this functionality will be handled on the experience page.